modules: add new module agg, renamed old aggregator to scalar_agg
`agg` is the module for aggregation of UniRec messages.  It can be
configured by user at startup (using cmd line arguments).
It is possible to set fields as a key of aggregation and to choose
aggregation functions and time interval.  The aggregated UniRec messages
are sent via output IFC.

`agg` was added into the RPM package for testing.

Co-authored-by: Michal Slabihoudek <slabimic@fit.cvut.cz>