urfilter: FEATURE subnets in IP comparison
It is now possible to check a subnet instead of just host IP.

The given subnet must be in form IP/BITS, where BITS is the number of
leading '1' bits in the netmask.

It should work both for IPv4 and IPv6.
All comparison operators should work but testing is needed.

Example filters:
-F 'DST_IP == 192.168.1.0/24'
-F 'DST_IP == 0.0.0.0/0'

Example output:

```
$ ./unirecfilter -F 'DST_IP == 192.168.1.0/3' -i f:data.trapcap,u:abc
Output interfaces: 1
[abc:buffer=off] Filter: DST_IP == 192.0.0.0/3 (224.0.0.0)
```

changelog: Add subnets feature to unirecfilter.