logreplay: fixed bug when logreplay did not terminate
Logreplay did not properly terminate if there was connected a client
which was not receiving. The behavior was changed to:

First SIGINT: attempt to break the main loop and send EOF message to the connected client.
Second SIGINT: forcefully terminate libtrap and thus don't send EOF message to the connected client (useful if the client is not responding).