merger: user-defined output template
Avoid waiting for the first incoming message when user set output template.
As @qha suggested in
https://github.com/CESNET/Nemea-Modules/pull/132#issuecomment-518818114
it is not necessary to wait for input templates on the input IFCs when a
user supplies output template.

The behavior of merger was changed so the output format matches the user
template.  Therefore, it is no longer needed to receive the first
message in this mode.

When the output template is not passed, merger waits for input templates
on all input IFC and merges them to create output template.

Example:

```
./merger -i f:in0.trapcap,f:in1.trapcap,f:out:w -u "ipaddr SRC_IP,ipaddr DST_IP2,uint64 BYTES,uint64 BYTES2,time TIME_FIRST"
```