unirecfilter: "In Array" feature
This is a draft of an extension of liburfilter to support optimized
filtering of integer values like this:

`DST_PORT in [1, 234, 123, 80, 443]`

The array (in brackets `[` and `]`)is parsed, sorted, and finnaly in
evalAST() searched using bsearch().

This should be faster than complicated filter like this:

`DST_PORT == 1 or DST_PORT == 234 or DST_PORT == 123 or DST_PORT == 80 or DST_PORT == 443`

Note: this patch also adds a basic test in cmocka. To run it,
libcmocka-devel (or equivalent) must be installed.

The patch should be improved:
* distinguish signed / unsigned numbers (right now everything is
  compared as uint64_t)
* add missing UniRec types like ipaddr, macaddr, time