flow_meter: dns-sd: adopted code from dnsplugin
The adopted code was copied instead of referenced because the original
was not suited for use as a module.
As a consequence, three plugins now have duplicate code:
DNS plugin, Passive DNS plugin and the DNS-SD plugin.
After the completion of this plugin, an extraction of duplicate code
to a separate module is suggested.

As for code changes made in the adoption:
 - Function and variable names were changed to fit plugin name.
 - Export fields were changed to those from previous commit.
 - Return values in core functions were changed from FLOW_FLUSH to 0, as we want to aggregate as much as possible into single flow.
 - Most dns types in process_rdata() were trimmed, as are not needed.
 - filtered_append() calls were added to parser.
 - Debug messages were enabled and made more compact.
 - Minor codestyle changes.