Revert "build: gcc10+ compatibility: add -fcommon if gcc version is at least 10"
This reverts commit 09cd6b054b541f67d5354daf9c4a93f3a2d363e2 beacause
it is not universal as to detect GCC version it uses the -dumpversion
option.
According to manual:
> Depending on how the compiler has been configured it can be just
> a single number (major version), two numbers separated by a dot
> (major and minor version) or three numbers separated by dots
> (major, minor and patchlevel version).

If the reported version is not just the major version, the resulting
string cannot be parsed as an integer and the fix does not work.
This could be improved by:
> compiler_version="`$CC -dumpversion | cut -d. -f1`"
but there is a simpler way, so let's just revert the original fix and
try again.