unirecfilter: FEATURE host/port keywords in column name
For expressions, it is now possible to use `host` or `port`
instead of actual name of UniRec field. Internally, it is evaluated as a
comparison of `SRC_IP` and `DST_IP` for `host`, and `SRC_PORT` and
`DST_PORT` for `port`.

`host` can be used with IP or IP prefix as follows:
`host == 10.0.0.0/16` -> `SRC_IP == 10.0.0.0/16 or DST_IP == 10.0.0.0/16`
`host == 192.168.0.1` -> `SRC_IP == 192.168.0.1 or DST_IP == 192.168.0.1`

`port` can be used with unsigned int as follows:
`port < 200` -> `SRC_PORT < 200 or DST_PORT < 200`

Both keywords can be used with IN ARRAY function, such as:

`host IN [195.113.232.0/24, 147.32.0.0/16, 82.208.0.0/16] and port < 200 and not port in [80, 53, 123]`
`host in [10.1.1.1, 192.168.1.1, 192.168.2.1]`

Note: IP prefixes with IN ARRAY is also new feature added by this commit, but they should not be combined with IP addresses without prefix.