json_dump: Fixed output to stdout
The last commit (ab713e0), which added the option to stream data over TCP, broke the default output to TCP (stdout.write only accepts str, not bytes). Moreover, it used an undocumented socket.SocketIO object, which is not the best idea.

This implements the TCP stream differently and reverts the changes done to the other output options, so all output types now work.